Revert "Refactor folder so it's more understandable for onboarding team members"
