+++ src/ThemeProvider.js
@@ -2,9 +2,9 @@
 
 import * as React from 'react';
 import { ThemeProvider } from 'styled-components';
-import defaultTheme from '../';
+import defaultTheme from './themes';
 import type { Props } from './ThemeProviderTypes';
-import './baseStyle';
+import './base.styled';
 
 const PhoenixThemeProvider = (props: Props) => {
   const theme: Object = { ...defaultTheme, ...props.theme };
