+++ src/example/example.test.js
@@ -1,0 +1,9 @@
+// @flow
+
+/* eslint flowtype-errors/enforce-min-coverage: 0 */
+
+describe('Example Test', () => {
+  it('Just testing', () => {
+    expect(1).toBe(1);
+  });
+});
