--- src/themes/ThemeProvider/index.js
@@ -1,7 +1,0 @@
-// @flow
-
-/* eslint flowtype-errors/enforce-min-coverage: 0 */
-
-import ThemeProvider from './ThemeProvider';
-
-export default ThemeProvider;
